Move ClassPathUtils to querydsl-codegen with required dependencies. Restore Reflection API dependency scope to test.  Change package names in troubleshooting.xml.
