+++ querydsl-codegen/pom.xml
@@ -39,6 +39,12 @@
       <artifactId>javax.inject</artifactId>  
       <version>1</version>  
     </dependency>
+
+    <dependency>
+      <groupId>org.reflections</groupId>
+      <artifactId>reflections</artifactId>
+      <version>0.9.9</version>
+    </dependency>
     
     <!-- test -->    
     <dependency>
