+++ querydsl-codegen/src/test/java/com/querydsl/SomeOtherClass.java
@@ -1,9 +1,9 @@
-package com.querydsl.core.util;
+package com.querydsl;
 
 public class SomeOtherClass {
 
     static {
-        SomeOtherClass2.property = "XXX";
+      SomeOtherClass2.property = "XXX";
     }
     
     public String someProperty;
