+++ querydsl-codegen/src/test/java/com/querydsl/SomeOtherClass2.java
@@ -1,4 +1,4 @@
-package com.querydsl.core.util;
+package com.querydsl;
 
 public class SomeOtherClass2 {
